world_model/contrasive_experience/requirements_web.txt
accelerate==1.11.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiolimiter==1.2.1
aiosignal==1.4.0
aiosqlite==0.21.0
alphashape==1.3.1
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
beartype==0.22.5
beautifulsoup4==4.14.2
Brotli==1.1.0
bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
click-log==0.4.0
clip==0.2.0
cloudpickle==3.1.2
contourpy==1.3.2
Crawl4AI==0.7.6
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
dashscope==1.24.9
datasets==4.3.0
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
einops==0.8.1
eval_type_backport==0.2.2
evaluate==0.4.6
exceptiongroup==1.3.0
faiss-cpu==1.12.0
fake-http-header==0.3.5
fake-useragent==2.2.0
Farama-Notifications==0.0.4
fastuuid==0.14.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
gymnasium==1.2.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
json5==0.12.1
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.3.1
litellm==1.79.1
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==2.7.0
packaging==25.0
pandas==2.3.3
patchright==1.55.2
peft==0.17.1
pillow==12.0.0
platformdirs==4.5.0
playwright==1.55.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qwen-agent==0.0.31
qwen-vl-utils==0.0.14
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
rtree==1.4.1
safetensors==0.6.2
scipy==1.15.3
sentry-sdk==2.43.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8
sympy==1.14.0
text-generation==0.7.0
tf-playwright-stealth==1.2.0
tiktoken==0.12.0
tokenizers==0.21.4
tqdm==4.67.1
transformers==4.49.0
trimesh==4.9.0
triton==3.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.22.3
websocket-client==1.9.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
